retourne pas de int

+++ c-skeleton/headers/assign_vector_to_centro.h
@@ -5,6 +5,6 @@
 #include "point.h"
 #include "distance.h" 
 
-int assign_vector_to_centroids(cluster_t* centroids, cluster_t* clusters, uint32_t K, squared_distance_func_t distance_func);
+uint64_t assign_vector_to_centroids(cluster_t* centroids, cluster_t* clusters, uint32_t K, squared_distance_func_t distance_func);
 
 #endif /* ASSIGN_VECTORS_TO_CENTROIDS_H */
